scan from the transaction local data once we've ran through all the available manifest data files
