+++ CMakeLists.txt
@@ -45,6 +45,7 @@ set(EXTENSION_SOURCES
     src/storage/authorization/sigv4.cpp
     src/storage/authorization/none.cpp
     src/storage/authorization/oauth2.cpp
+    src/storage/iceberg_transaction_data.cpp
     src/storage/irc_authorization.cpp
     src/storage/irc_catalog.cpp
     src/storage/irc_schema_entry.cpp
