+++ src/iceberg_functions/iceberg_multi_file_list.cpp
@@ -40,6 +40,15 @@ const IcebergTableMetadata &IcebergMultiFileList::GetMetadata() const {
 	return scan_info->metadata;
 }
 
+bool IcebergMultiFileList::HasTransactionData() const {
+	return scan_info->transaction_data;
+}
+
+const IcebergTransactionData &IcebergMultiFileList::GetTransactionData() const {
+	D_ASSERT(HasTransactionData());
+	return *scan_info->transaction_data;
+}
+
 optional_ptr<IcebergSnapshot> IcebergMultiFileList::GetSnapshot() const {
 	return scan_info->snapshot;
 }
@@ -314,27 +323,32 @@ optional_ptr<const IcebergManifestEntry> IcebergMultiFileList::GetDataFile(idx_t
 
 	while (file_id >= data_files.size()) {
 		if (current_data_files.empty() || data_file_idx >= current_data_files.size()) {
+			current_data_files.clear();
 			//! Load the next manifest file
-			if (current_data_manifest == data_manifests.end()) {
+			if (current_data_manifest != data_manifests.end()) {
+				auto &manifest = *current_data_manifest;
+				auto full_path = options.allow_moved_paths ? IcebergUtils::GetFullPath(path, manifest.manifest_path, fs)
+				                                           : manifest.manifest_path;
+				auto scan = make_uniq<AvroScan>("IcebergManifest", context, full_path);
+
+				data_manifest_reader->Initialize(std::move(scan));
+				data_manifest_reader->SetSequenceNumber(manifest.sequence_number);
+				data_manifest_reader->SetPartitionSpecID(manifest.partition_spec_id);
+
+				while (!data_manifest_reader->Finished()) {
+					data_manifest_reader->Read(STANDARD_VECTOR_SIZE, current_data_files);
+				}
+				current_data_manifest++;
+			} else if (HasTransactionData() && !scanned_transaction_data) {
+				auto &transaction_data = GetTransactionData();
+				auto &data_files = transaction_data.manifest_file.data_files;
+				current_data_files.insert(current_data_files.end(), data_files.begin(), data_files.end());
+				scanned_transaction_data = true;
+			} else {
 				//! No more data manifests to explore
 				return nullptr;
 			}
 
-			auto &manifest = *current_data_manifest;
-			auto full_path = options.allow_moved_paths ? IcebergUtils::GetFullPath(path, manifest.manifest_path, fs)
-			                                           : manifest.manifest_path;
-			auto scan = make_uniq<AvroScan>("IcebergManifest", context, full_path);
-
-			data_manifest_reader->Initialize(std::move(scan));
-			data_manifest_reader->SetSequenceNumber(manifest.sequence_number);
-			data_manifest_reader->SetPartitionSpecID(manifest.partition_spec_id);
-
-			current_data_files.clear();
-			while (!data_manifest_reader->Finished()) {
-				data_manifest_reader->Read(STANDARD_VECTOR_SIZE, current_data_files);
-			}
-
-			current_data_manifest++;
 			data_file_idx = 0;
 		}
 
@@ -367,10 +381,6 @@ OpenFileInfo IcebergMultiFileList::GetFile(idx_t file_id) {
 	lock_guard<mutex> guard(lock);
 	if (!initialized) {
 		InitializeFiles(guard);
-	}
-
-	if (!scan_info->snapshot) {
-		return OpenFileInfo();
 	}
 
 	auto found_data_file = GetDataFile(file_id);
