+++ src/include/storage/irc_table_set.hpp
@@ -2,6 +2,7 @@
 #pragma once
 
 #include "storage/irc_table_entry.hpp"
+#include "storage/iceberg_transaction_data.hpp"
 
 namespace duckdb {
 struct CreateTableInfo;
@@ -12,21 +13,6 @@ class IRCTransaction;
 struct IRCAPITableCredentials {
 	unique_ptr<CreateSecretInput> config;
 	vector<CreateSecretInput> storage_credentials;
-};
-
-struct IcebergTransactionData {
-public:
-	IcebergTransactionData(IcebergManifestFile &&manifest_file, IcebergManifestList &&manifest_list,
-	                       IcebergSnapshot &&snapshot)
-	    : manifest_file(std::move(manifest_file)), manifest_list(std::move(manifest_list)),
-	      snapshot(std::move(snapshot)) {
-	}
-	static unique_ptr<IcebergTransactionData> Create(ClientContext &context, IcebergTableInformation &table_info);
-
-public:
-	IcebergManifestFile manifest_file;
-	IcebergManifestList manifest_list;
-	IcebergSnapshot snapshot;
 };
 
 struct IcebergTableInformation {
